Refactoring: Update VerifiedPassword and VerifiedUsername validators
