+++ backend/monolith/mono-security-service/src/main/java/dev/alexengrig/myim/mono/security/validation/VerifiedUsername.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.alexengrig.myim.mono.security.validation.username;
+package dev.alexengrig.myim.mono.security.validation;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
